+++ front_end/src/components/Trading/Orders.tsx
@@ -12,7 +12,7 @@ import { Container } from 'react-bootstrap'
 import { useDispatch, useSelector } from 'react-redux'
 import { type Order, type tradingDataDef } from '../DataManagement'
 import { filterSlice, type FilterState } from '../StateManagement'
-import './tables.css'
+import '../../css/charts.css'
 
 interface TableProps {
   orders: Order[]
