+++ front_end/src/css/tables.css
@@ -3,6 +3,7 @@
   --ag-grid-size: 5px;
   --ag-list-item-height: 20px;
   --ag-selected-row-background-color: rgb(0, 255, 0, 0.1);
+  --ag-background-color: rgb(0, 0, 0, 0.1);
 }
 
 .buy-cell {
